requirements.txt
### AI & LLM:
llama-index==0.12.23
llama-index-readers-json==0.3.0
google-genai==1.11.0
pandas==2.2.3
### Web/API:
fastapi==0.112.2
uvicorn
python-multipart==0.0.20
Jinja2
wikipedia==1.4.0
### Database & Task Queue:
sqlmodel==0.0.24
asgiref==3.8.1
psycopg2-binary==2.9.10
alembic==1.15.2
celery==5.5.2
redis==5.2.1
boto3==1.38.8
### File/Document Handling:
PyMuPDF==1.25.5
html2text==2024.2.26
openpyxl==3.1.5
python-docx==1.1.2
xlrd==2.0.1
extract-msg==0.54.1
markdownify==1.1.0
magika==0.6.1
### Audio/Voice:
SpeechRecognition==3.14.2
pydub==0.25.1
### Utility & Others:
python-dotenv==1.0.1
cryptography
theflow==0.8.6
geopy==2.4.1